requirements.txt
Django==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
requests==2.31.0
